requirements2.txt
absl-py==0.15.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==21.3.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cycler==0.11.0
dataclasses==0.6
decorator==5.1.0
defusedxml==0.7.1
dm-tree==0.1.6
entrypoints==0.3
filelock==3.4.1
flatbuffers==1.12
gast==0.3.3
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jupyterthemes==0.20.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tuner==1.1.0
kiwisolver==1.3.1
kt-legacy==1.0.4
lesscpy==0.15.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.22.2
oauthlib==3.1.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
ply==3.11
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.15.0
soupsieve==2.3.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.16.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-model-optimization==0.7.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.12.1
zipp==3.6.0
